+++ WasapiIODLL/WasapiIOIF.cpp
@@ -227,7 +227,7 @@ WasapiIO_AddPlayPcmDataStart(void)
 
 extern "C" __declspec(dllexport)
 bool __stdcall
-WasapiIO_AddPlayPcmData(int id, unsigned char *data, int bytes)
+WasapiIO_AddPlayPcmData(int id, unsigned char *data, int64_t bytes)
 {
     assert(self);
     return self->playPcmGroup.AddPlayPcmData(id, data, bytes);
@@ -235,7 +235,7 @@ WasapiIO_AddPlayPcmData(int id, unsigned char *data, int bytes)
 
 extern "C" __declspec(dllexport)
 bool __stdcall
-WasapiIO_AddPlayPcmDataSetPcmFragment(int id, int posBytes, unsigned char *data, int bytes)
+WasapiIO_AddPlayPcmDataSetPcmFragment(int id, int64_t posBytes, unsigned char *data, int64_t bytes)
 {
     assert(self);
 
@@ -316,22 +316,22 @@ WasapiIO_SetNowPlayingPcmDataId(int id)
 
 extern "C" __declspec(dllexport)
 void __stdcall
-WasapiIO_SetupCaptureBuffer(int bytes)
+WasapiIO_SetupCaptureBuffer(int64_t bytes)
 {
     assert(self);
     self->wasapi.SetupCaptureBuffer(bytes);
 }
 
 extern "C" __declspec(dllexport)
-int __stdcall
-WasapiIO_GetCapturedData(unsigned char *data, int bytes)
+int64_t __stdcall
+WasapiIO_GetCapturedData(unsigned char *data, int64_t bytes)
 {
     assert(self);
     return self->wasapi.GetCapturedData(data, bytes);
 }
 
 extern "C" __declspec(dllexport)
-int __stdcall
+int64_t __stdcall
 WasapiIO_GetCaptureGlitchCount(void)
 {
     assert(self);
@@ -380,7 +380,7 @@ WasapiIO_Unpause(void)
 }
 
 extern "C" __declspec(dllexport)
-int __stdcall
+int64_t __stdcall
 WasapiIO_GetPosFrame(void)
 {
     assert(self);
@@ -388,7 +388,7 @@ WasapiIO_GetPosFrame(void)
 }
 
 extern "C" __declspec(dllexport)
-int __stdcall
+int64_t __stdcall
 WasapiIO_GetTotalFrameNum(void)
 {
     assert(self);
